Move `CompositeGroup` to separate file
